Initial PoC
Remove global CSS settings from tableToSVG.css, as they pertained only
to the documentation maintained by the lib owner.

Create loanauction.js, which injects an SVG visualisation of the bid
summary table by making a copy and scrubbing it so that createChart can
safely parse it.

Create a Chrome extension manifest that injects loanauction.js along
with the resources in table-to-svg for use in Funding Circle auction
pages.

TODO: Trim bid sub-groups from the graph, handle re-rendering of graph
when bid-summary is wiped and repopulated