Upgrade to content injection
inject the scripts of interest into the loan auction page itself so that
they have access to the page globals, notably a function that populates
the panes used for placing bids.

Modify loanauction.js to reflect the change in approach, notably
introducing an fcViz callback. Bind this to the change event on tabs.

TODO: work out one more corner case to ensure that the chart will be
there.