Directly use D3
Use an array of tuples instead of two arrays to hold the data to make it
easier to funnel them into D3.
Follow the tutorial given in D3 for bar charts, making slight
adjustments to account for the fact that we have to manually inject the
svg element into the DOM.
Remove all logic that was previously used to prepare data for
tableToSVG's createChart.

TODO: Replace x-axis label clutter with tooltips on the bars, use more
concise y-axis labelling, work out how to indicate where user's bids are
in the summary.