Add tooltips, tweak appearance
Add tooltips to the bar charts indicating how much was bid at what rate.

Make the cumulative line chart 50% transparent, making it fully visible
on mouseover. Thicken the line to compensate for reduction in
visibility, and use a colour found on the page.

TODO: plotting of user bids on the chart, consider populating the bid
rate when the user clicks on the bar