All Bids pagination refactoring
refactor renderAllBidCharts so that it polls all pages sequentially,
pushing each response into an array.

refactor makeAllDataFrom to parse objects, and change the data structure
to be a nested map of maps, keyed first by (Date,Rate) and then by user.
keep an array of the keys within each level of the map