Insert scatter plot infobox
Inject a div that will hold details from a selected point in the scatter
plot. Adjust the chart dimensions and hint text so that we can allocate
as much real estate to the infobox while minimising disruptions to
layout.

Note that this infobox currently obscures the live auction countdown
timer.