Generate a donut chart
Ensure that when the close button is clicked on the infobox that we
forget window.clickedKey.

Generate a pie chart using the user totals obtained from dataBlob[key],
where key is a time-rate tuple. Maintain a mapping of users to colors,
generating one for a user if not found in the mapping.