Final touches to all bids
Adjust the bid infobox width so that they don't completely cover the
live auction countdown, which will matter during the final hour of the
auction

Show the amount bid and rate in the middle of the donut

When the parts of the donut are mouseovered, show only the bids that
make up that part in the bids table, thereby identifying the user and
the earliest bids he made. Change the donut text so that it shows how
much he bid.

This concludes the initial implementation of the all bids chart scatter
plot.