Calculate gmtOffset
The displayTime is usually displayed in either GMT or BST, while the
created_at field is always given in UTC milliseconds. Work out the
offset between the two, and use that to adjust created_at. This will
lead to a chart using a timezone independent of the one the client
machine is using.