Introduce enhancements to requests
Adjust loanauction.js so that its mutation observer can be inactivated,
and that it does not call the chart rendering callback if activeId is
undefined, ie, if we are not in a loan auction page, which contains the
bid tabs.

Build out an injector script for loan requests and add it to the
manifest. This will inject both d3 and loanauction since we might need
those to render a bid summary chart in the requests page. Add to these a
rudimentary requests.js script that currently just prepends the request
titles with a text-based link that can be clicked to expand the entry
for more content.