Build out loan request preview pane
Use span instead of a tags to hold the text to expand/collapse the
preview panes. This should make it more pleasant to look at without the
underline. Adjust the font accordingly and move the CSS pointer style
from .js to .css.

Instead of using a switch statement to work out what to do on clicking,
just work out if we have a preview pane available to use, creating one
if we don't, and use jQuery.animate to animate both the expansion and
collapse of preview element.

Create both a filler and a preview pane so that in creating two tr
elements we don't disrupt the alternating row colour scheme in the
table.