Add company profile to preview pane
Almost double the height for preview pane in CSS.

Add explicit dimension and padding to the right pane so that it plays
nicely when rendered as a scroll-box.

Grab the company profile from the auction response and tailor it to fit
into the pane. In particular, remove the float attribute so that it will
flow around the non-limited company credit score table, which has now
been made to float right.