Restructure repayByDate, flesh out chart
Restructure repayByDate such that the values held include by-date
principal/interest/fee numbers.
Use this for the graph that we are generating Funding Circle.
Adjust the y axis domain so that we can actually see the bars. We should
consider running the bars and cumulative cashflow on separate axes.
Note that fees are not shown, as they have been found to be too
minuscule, and I have yet to work out how to display negative bars.
We probably need to work out how to introduce chart scrolling/zooming
such that users wanting to see subsequent years can scroll to the right.
We can then make the bar charts fatter.