Add c3, refactor injectors
Add c3 as a dependency. This will give us charting capabilities that are
suitable for the repayment schedule graph that we cannot do with d3
alone.

Refactor the injector scripts to improve readability. We now call a
recursive chainLoad function with an array of resource paths that we
want chain loaded. chainLoad calls an inject factory function, supplying
as arguments the resource to be loaded in a given recursion step and
either a call to itself, or null, depending on whether there are any
resources left to be loaded.