Property investor reports in financial summary
Update the manifest to allow access to the webRequest API for the
Funding Circle download URLs

Add a background script that listens for responses to web requests for
PDFs from Funding Circle. Modify the Content-Type and
Content-Disposition headers for responses to URLs anchored with #fcviz,
so that we are not forced to save such URLs to disk.

loanauction.js: update fcVizObserver to inspect mutation.target.id. For
financial summary targets, if the PDF hyperlink starts with Investor
Report, we can safely assume that we are dealing with a property loan.

When so, detach the Key Financials table which contains no information,
and replace with an iframe whose source is an #fcviz-anchored PDF URL
derived from the hyperlink mentioned earlier.

TODO: Cleanup the layout.