Add bid peeking to watchlist/loan requests
Attach an onClick handler to all auction hammer images, which will send
an http request to the auction's my bids page.
On success, extract the live and rejected bid amounts from the response
an append to the Amount cell.

TODO: Busy indicator, tooltip to display full response or help message
(to click on hammer)