Add bid details to auction tooltip when mousing over
Retain all responses from my_bids in a global map keyed by id, trimming
off the lender name. On mouseover, if we have an entry in myBids,
populate a tooltip with that information and show it near the mouse
pointer. Otherwise, just fill it in with a hint to the user to click on
the icon to get the bids.

TODO: auction icon to repeatedly fade in and out to indicate busyness