Properly target individual loans page
Realise that the individual loan parts tab is being held in an iframe, and requires corresponding changes to the extension manifest to properly target it.

TODO: Inject a new method into indiv parts controller that updates markup for all loan parts to be sold. Impl will be similar to how it's done for indiv parts, with changes to reflect the fact that it's done for multiple parts