Wire updateSalesInfoAll and clearSalesInfo
Declare a new function, findLoanPartFrom($event), that will look up a
loan part from the mouse event. $event will contain srcElement, so use
that to track down the table cell containing the id, and use that to
look up the loan part object.

Implement clearSaleInfo, which does the obvious.

Inject angular-compiled elements to the left and right of the markup
selector, which are bound to the clearSaleInfo and updateSaleInfoAll
handlers respectively.