Implement ajax-based filtering
Add a new pulser CSS class so that we can have an animated indicator for
activity being done in the background

Make changes to the form so that request-responses are handled using
ajax instead of regular http requests as it is now.

This may lead to a situation where the URL in the browser may not be
consistent with what is actually requested via the filter form. In light
of this, track the form requests made by the user and push the responses
and the associated url into browser history.

Further, if a user goes back to a page where all risk bands are
requested, fire a request to FC. This will reset FC's backend with the
form data associated with the page. This is needed because if the user
clicks on the paginator, we need to ensure that FC will not service the
request with stale form data from previous ajax queries.