Implement multi-band selection
Drop in chosen from http://harvesthq.github.io/chosen/ , and update
manifest to inject this in the requests page.

chosen-ise the risk band filter, and implement behaviour so that the All
and band-specific options are mutually exclusive.

Augment submit behaviour so that we fire off multiple request queries,
one per risk band, and collate the responses. Once we have them all, put
all the tbody rows into one section, such that the bands appear in the
sequence found in the form, taking care also to correctly display the
number of auctions found in total.

TODO:
- server-side is holding state, so we need to work out what to do when
the user:
- clicks away from the loan requests page and then back again
- clicks refresh
- implement client-side filtering
- implement ajax-based watchlist updates