Fix up multi-band filtering
Extract success handler for risk band filter

Make update watchlist ajax-ready, so that it does not reload the page.
This is needed if the page is displaying more than one risk band,
otherwise, the server-side will return only the resultset for one of the
bands.

If the user is going back and forth in history, ensure that we not only
have the e.state, but also the corresponding sections object held in
window.sections. Also, if the history object is associated with one risk
band, fire an ajax request to keep the server-side up to date.

Finally, store the original sections in window.sections, not clones, so
that we retain whatever chosen has prepared for those sections.