Refactor fcVizObserver, rearrange Term Sheet
Instead of making fcVizObserver a god function whose behaviour is driven
by a giant switch statement, make it simply invoke a callback, and let
each element of interest supply a callback specific to itself.

Through this, introduce a new repayments callback that will put the term
sheet on top of the comments list.