Make property layout avail to other pages
Extract triggerPropertyLayout to its own module, allowing us to inject
it into other pages.

Notice that the finance-summary pages exists, and are accessible from a
link to users on the loan parts page in the financial summary section.
Create a new set of entries in the manifest that cater to it.

TODO: Apply this to the financial summary pages in the loan parts page.