Render parts grid on-the-fly
* Given that many investors will have lots of loan parts, it makes more
sense for DataTable to render the parts on the fly. Rework the code to
do so.

* Because DataTable cannot render jQuery elements when given verbatim,
we have to supply them as raw HTML. This unfortunately means we lose
some event handlers for the Status cell data, ie, the comment mouseover
tooltips, but this shouldn't be a problem given that we can view them
from somewhere else

TODO: Group-by loan part, ensure we get the live loan parts first