Block FC's new web analytics script
The new script put in by FC records all mouse events, not just clicks on
specific elements. Given that FCViz makes significant changes to the
layout, any data recorded on such an augmented page will certainly be
inaccurate. Prevent the loading of the script so that we avoid tainting
FC's analytics